requirements.txt
Django==4.1.7
python-dotenv
mysqlclient